+++ packages/eslint-local-rules/src/packageJsonIconPaths.ts
@@ -87,7 +87,7 @@ export const packageJsonIconPaths: Rule.RuleModule = {
     }
   },
   create: context => {
-    const filename = context.filename ?? context.getFilename();
+    const filename = context.filename;
     if (!filename.match(/packages\/[^/]+\/package\.json$/)) {
       return {};
     }