+++ packages/playwright-vscode-ext/src/index.ts
@@ -30,7 +30,8 @@ export {
   ensureSecondarySideBarHidden,
   waitForExtensionsActivated,
   closeWorkspaceToEmptyWindow,
-  prepareNoFolderOpenForPaletteTests
+  prepareNoFolderOpenForPaletteTests,
+  disableMonacoAutoClosing
 } from './utils/workflows';
 
 export { activeQuickInputWidget, activeQuickInputTextField } from './utils/quickInput';
@@ -85,6 +86,31 @@ export {
 } from './pages/commands';
 export type { OpenCommandPaletteOptions } from './pages/commands';
 
+export {
+  focusOnFilesExplorer,
+  newUntitledTextFile,
+  saveFile,
+  clearAllNotifications,
+  closeAllEditors,
+  showExplorer,
+  reloadWindow,
+  goToFile,
+  goToLineColumn,
+  selectAll,
+  paste,
+  focusOnProblemsView,
+  goToDefinition,
+  focusActiveEditorGroup,
+  find,
+  hidePanel,
+  clearOutput,
+  closeEditor,
+  insertSnippet,
+  showRunningExtensions,
+  hideSecondarySideBar,
+  closeWorkspace
+} from './pages/nativeCommands';
+
 export { executeEditorContextMenuCommand, executeExplorerContextMenuCommand } from './pages/contextMenu';
 
 export {
@@ -121,6 +147,8 @@ export {
 
 export { webviewActiveFrame, hasTitle, hasContent } from './pages/webview';
 export type { ActiveFrameMatcher } from './pages/webview';
+
+export { TEST_EXPLORER_PANEL, TEST_EXPLORER_TREE_ITEM, focusAndTypeInFilter, clearFilter } from './pages/testExplorer';
 
 // Shared
 export { saveScreenshot } from './shared/screenshotUtils';